Create separate src file for AtlasManagerFilter
